Add functional coverage of connectoin_opts
The *_opts tests ensure that each function
exhibits the right behaviors in terms of how it populates
the config, but it does not verify that items are correctly
sourced if they're missing from knife.config.

These tests verify sourcing, and verify that options that
should not be sourced (like password) from Chef::Config
are not.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>